+++ packages/bottom-sheet/src/component.tsx
@@ -178,6 +178,7 @@ const TIMEOUT = 300;
 const SWIPE_CLOSE_VELOCITY = 0.4;
 const MIN_BACKDROP_OPACITY = 0.2;
 const HEADER_HEIGHT = 56;
+const MARKET_HEIGHT = 24;
 
 /* Верхний отступ шторки, если она открыта на максимальную высоту */
 export const HEADER_OFFSET = 24;
@@ -204,7 +205,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             titleAlign = 'left',
             trimTitle,
             stickyHeader,
-            stickyFooter,
+            stickyFooter = true,
             initialHeight = 'default',
             hideOverlay,
             hideHeader,
@@ -239,7 +240,6 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
             addonClassName,
             closerClassName,
             backerClassName,
-            swipeable,
             leftAddons,
             rightAddons,
             hasCloser,
@@ -274,7 +274,8 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
 
             if (
                 !scrollableContainer.current ||
-                (stickyHeader && offsetY <= HEADER_HEIGHT + HEADER_OFFSET)
+                (stickyHeader && offsetY <= HEADER_HEIGHT + HEADER_OFFSET) ||
+                (!stickyHeader && offsetY <= MARKET_HEIGHT + HEADER_OFFSET)
             ) {
                 return false;
             }
@@ -435,9 +436,7 @@ export const BottomSheet = forwardRef<HTMLDivElement, BottomSheetProps>(
                         })}
                         ref={scrollableContainer}
                     >
-                        {swipeable && (hideHeader || emptyHeader) && (
-                            <div className={cn(styles.marker)} />
-                        )}
+                        {swipeable && <div className={cn(styles.marker)} />}
 
                         {!hideHeader && !emptyHeader && <Header {...headerProps} />}