+++ packages/drawer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-drawer@3.1.0...@alfalab/core-components-drawer@3.1.1) (2022-03-24)
+
+**Note:** Version bump only for package @alfalab/core-components-drawer
+
+
+
+
+
 # [3.1.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-drawer@3.0.3...@alfalab/core-components-drawer@3.1.0) (2022-03-01)