+++ packages/gallery/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.5](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-gallery@3.1.4...@alfalab/core-components-gallery@3.1.5) (2022-03-28)
+
+**Note:** Version bump only for package @alfalab/core-components-gallery
+
+
+
+
+
+## [3.1.4](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-gallery@3.1.3...@alfalab/core-components-gallery@3.1.4) (2022-03-24)
+
+**Note:** Version bump only for package @alfalab/core-components-gallery
+
+
+
+
+
 ## [3.1.3](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-gallery@3.1.2...@alfalab/core-components-gallery@3.1.3) (2022-03-22)
 
 **Note:** Version bump only for package @alfalab/core-components-gallery