+++ packages/tooltip/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.1](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-tooltip@5.0.0...@alfalab/core-components-tooltip@5.0.1) (2022-03-28)
+
+**Note:** Version bump only for package @alfalab/core-components-tooltip
+
+
+
+
+
+# [5.0.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-tooltip@4.6.0...@alfalab/core-components-tooltip@5.0.0) (2022-03-24)
+
+
+### Features
+
+* **bottom-sheet:** update-bottom-sheet ([#1025](https://github.com/alfa-laboratory/core-components/issues/1025)) ([26fa9aa](https://github.com/alfa-laboratory/core-components/commit/26fa9aab68bebf0f7093a38bc0f18a9b596ccf37)), closes [#1032](https://github.com/alfa-laboratory/core-components/issues/1032)
+
+
+### BREAKING CHANGES
+
+* **bottom-sheet:** Большое обновление стилей, множество дополнительных настроек
+
+
+
+
+
 # [4.6.0](https://github.com/alfa-laboratory/core-components/compare/@alfalab/core-components-tooltip@4.5.1...@alfalab/core-components-tooltip@4.6.0) (2022-03-22)